# hackathonApp/backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon App API
  description: |
    REST API for the Hackathon Application. Supports user authentication (with optional
    MFA), ticket management, real-time chat (channels, DMs, ticket threads), Workday
    ad-hoc sync for admin users, and MFA administration.

    ## Authentication
    All authenticated endpoints require a valid session cookie (`hackathon_session`).
    Obtain a session by calling `POST /api/auth/login`. If the user has MFA enabled,
    a two-step login flow is required (see MFA section).

    ## Rate Limiting
    All endpoints are rate-limited. Unauthenticated endpoints are limited per IP;
    authenticated endpoints are limited per user. When rate-limited, the API returns
    `429 Too Many Requests` with a `Retry-After` header.

    ## Error Format
    All errors follow a consistent JSON format:
    ```json
    {
      "error": "error_code_or_message",
      "message": "Human-readable description"
    }
    ```
  version: 1.0.0
  contact:
    name: Hackathon App Team
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://hackathon.example.com/api
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: |
        Authenticates a user. If MFA is enabled for the user, returns an `mfa_token`
        instead of a full session. The client must then call `/auth/mfa/verify-login`
        with the token and a TOTP code to complete login.

        If a valid `device_token` cookie is present (trusted device), MFA is skipped.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful (no MFA) or MFA required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LoginSuccess'
                  - $ref: '#/components/schemas/MFARequired'
        '401':
          description: Invalid credentials
        '429':
          description: Rate limited (5 per minute per IP)
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
      responses:
        '201':
          description: Registration successful
        '400':
          description: Validation error (weak password, duplicate email)
        '429':
          description: Rate limited (3 per minute per IP)
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and clear session
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Logged out
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh session to extend expiration
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Session refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccess'
        '401':
          description: Session expired
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated

  # ──────────────────────────────────────────────
  # MFA
  # ──────────────────────────────────────────────
...
  /admin/sync/user/{workday_id}:
    post:
      tags: [Admin - Sync]
      summary: Sync a single user from Workday
      security:
        - sessionCookie: []
      parameters:
        - name: workday_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                  worker:
                    $ref: '#/components/schemas/User'
                  changed_fields:
                    type: array
                    items:
                      type: string
                  previous_values:
                    type: object
        '403':
          description: Admin access required
        '503':
          description: Workday unavailable, cached data returned
  /admin/sync/bulk:
    post:
      tags: [Admin - Sync]
      summary: Trigger bulk sync of all users
      security:
        - sessionCookie: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                filter:
                  type: object
                  properties:
                    team_name:
                      type: string
                    table:
                      type: string
                force:
                  type: boolean
                  default: false
      responses:
        '202':
          description: Bulk sync started
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                  total_workers:
                    type: integer
                  status_url:
                    type: string
        '403':
          description: Admin access required
        '429':
          description: Bulk sync already in progress
  /admin/sync/status/{job_id}:
    get:
      tags: [Admin - Sync]
      summary: Check sync job status
      security:
        - sessionCookie: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status with progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobStatus'
        '404':
          description: Job not found
  /admin/sync/history:
    get:
      tags: [Admin - Sync]
      summary: Get sync audit history
      security:
        - sessionCookie: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 25
        - name: trigger_type
          in: query
          schema:
            type: string
            enum: [manual, bulk, first_login, stale_login]
        - name: status
          in: query
          schema:
            type: string
            enum: [completed, failed, partial, running]
        - name: since
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated sync history

  # ──────────────────────────────────────────────
  # ADMIN - MFA
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hackathon_session
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [hacker, expert, admin]
        workday_id:
          type: string
        department:
          type: string
        table:
          type: string
        team_name:
          type: string
        is_stale:
          type: boolean
        last_synced_at:
          type: string
          format: date-time
          nullable: true
    LoginSuccess:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
    MFARequired:
      type: object
      properties:
        mfa_required:
          type: boolean
          example: true
        mfa_token:
          type: string
        mfa_token_expires:
          type: integer
          description: Seconds until mfa_token expires
...
    SyncJobStatus:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, partial]
        trigger_type:
          type: string
        initiated_by:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            name:
              type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        progress:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
            succeeded:
              type: integer
            failed:
              type: integer
            skipped:
              type: integer
            percent_complete:
              type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              workday_id:
                type: string
              error:
                type: string